split markup header files
I found it very hard to find my way around in the markup namespace. Of
course it's a matter of taste whether headers should be topic centered
or one-class-one-header…
Also, initially it was confusing that this is completely unrelated to
markup parsing so it seems reasonable to put this into the output
namespace.