do not bundle externals but require them as dependencies
The process of throwing the externals out is not complete yet but we
already updated the LICENSE to reflect that.

To make this fully feasible we need a way to tell developers how they
can get a hold of all these dependencies. We would propose to use conda
as things can easily be packaged there for Linux/Windows/macOS which is
should cover virtually all developers out there. But not all
dependencies are packaged yet. In particular we are waiting for cppast
to have a release.