use model of value type
Replaced the somewhat rigid tree of markup entities with a tree of value
types. By not using pointers anywhere anymore this hopefully makes it
much easier to transform this tree. This also splits up all the steps to
build that final tree into individual somewhat minimal transformations
that are hopefully easier to understand.