prefer the term (link) target over the confusing term anchor
in the HTML standard there are two types of anchors: origin anchors and
destination anchors. While the `<a>` is almost exclusively used for
origins, i.e., to actually create a clickable hyperlink, most people
actually mean the destination when they talk about an anchor, which is
however usually encoded with an `id` attribute.

To avoid this confusion, we use the term `link` and `target` instead.